move logs to a new page
