+++ src/app/logs/components/LogFooter.tsx
